vql/parsers/sdjournal: Add plugin to read systemd journal on Linux
Newer Linux systems use the systemd journal facility for logging instead
of the classic syslogd logging daemon.  While it's possible to configure
journald to forward events to e.g. rsyslogd, doing so requires a
configuration change on the endpoint that would only be necessary to
accommodate the syslog parser.

This commit adds the ability to parse the systemd journal using the
native libsystem-journal.so shared library.  It adds watcher and scanner
parsers identical to the syslog versions.  One difference is that since
the underlying library handles opening and tracking the journal files,
there is no facility to subscribe to specific journal files.  The
support is also limited to the system journal, but that should
be sufficient for Velociraptor.  It is expected that queries against
these parsers filter on systemd unit (or other fields)
as desired rather than implementing those filters in the parser itself.