vql/linux/tcpsnoop: Add plugin to report incoming/outgoing TCP connections
The plugin utilizes the ebpf infrastructure to load a small program
which hooks into tcp_connect/tcp_accept code paths of the kernel and
print information about the connection source/dest ports/ip addresses,
PID/UID of the process initiating the connection, ip protocol version,
timestamp, and the direction of the connection.

libbpfgo is added as a git submodule that will only be imported and
patched if required.  The patching of libbpfgo is limited to statically
linking libelf.

In order to build it you'll need several things:
 * At least version 10 of clang compiler, preferably 13
 * llvm-strip tool

Then you can run the build like so:

$ make linux

If you would like to skip bpf support, you can build with
$ make linux BUILD_LIBBPFGO=0

It's important to run the "linux" target from make as opposed to simply
running "make" as the latter would only run the build process from mage.