vql/linux/chattrsnoop: add plugin and artifact for detecting immutable files
It hooks into do_vfs_ioctl and also performs some of the checks that the
generic code would perform in order to ensure we are triggering the ebpf
kprobe on a fs which indeed supports the given operations.

While at it remove the vmlinux.h files that comes with tcpsnoop and
instead rely on the one provided by libbpf.

Add artifact for chattrsnoop plugin