api: add type and description fields to v1/GetArtifacts endpoint
The v1/GetArtifacts endpoint only allows entire artifacts or just the
name to be returned.  In order to make the UI for managing client events
nicer, we'd like to be able to search and present a set of switches to
select the client artifacts.  This change adds the ability to request
the type and description fields.