vql/linux/audit: rearchitect plugin for scalability
The audit plugin doesn't share anything among invocations, which makes
sense given the historical model of preconfiguring of the rules,
subscribing to the firehose, and sorting out the events later.  The problem
with that architecture is that if anything external happens with the
audit subsystem -- even as simple as restarting the audit daemon -- the
audit rules are lost and no more events will be generated.

By specifying which rules a particular artifact depends on in the
arguments, we can be certain that the rules required will always be in
effect, can log when they are not, and restore them automatically.

Since multiple invocations can ultimately require the same audit rules,
management of those rules must be centralized and reference counted so
they are only removed when the last plugin using a rule exits.

There are a few changes required to make this go:
- We need a separate command-and-control audit client to do the
  configuration, checking, and fixing
- We need a goroutine that periodically checks to ensure that the rules
  are correct.  This is implemented as something that sits idle waiting
  for configuration events and then batches them until 500ms has passed
  since receipt of the last one.  The reason for this is that e.g.
  "auditctl -D" will generate an event for every single rule removed
  and we don't need to check the rules every time.
- Each invocation will receive a copy of each event and also log
  messages that apply to the invocation as needed.
- The lifetime of the shared services is tied to the listeners.
  Once the last listener exits, the shared services shutdown.  If new
  invocations come in, a new instance will be started up.
- In order to not conflict with auditd, we must use the multicast
  interface for receiving messages.  The kernel does *not* queue and
  retry multicast audit messages so we need to be very careful to avoid
  dropping events.
- In order to avoid dropping events on very busy systems, we increase
  the default socket receive buffer to 512k and allow it to scale up
  as needed in response to ENOBUFS from the kernel.
- We use a separate buffer for each event received so that the
  reception of the messages and processing of events can be decoupled.
- These buffers are kept in a pool to avoid constant allocation and
  freeing.
- We use a FileBytesRingBuffer to queue events so that we don't end up pinning
  lots of memory if events get backlogged.
- To aid in debugging and monitoring, there are statistics gathered.  A
  global variable can be used to enable printing the statistics to the
  client terminal every 5 seconds.