api: add with_roles to v1/GetUsers to return roles for each user
The coming user admin GUI has the ability to filter users by name and
by role assignment.  While we could fetch a record for each user to
obtain the roles, it's much more efficient to provide the roles with
the listing.  Since v1/GetUsers is now only available to users with
the SERVER_ADMIN role and the user list is filtered for users
without the ORG_ADMIN role, there is no information leakage.