api: add v1/CreateUser endpoint
The user admin GUI needs the ability to create a user.  This endpoint
will accept a POST request to create a user or edit an existing one,
depending on whether the Replace flag is set.  It requires the
SERVER_ADMIN role to use and membership in any organizations the request
contains (or ORG_ADMIN role).