gui: add user admin tool
This commit implements the GUI portion of the user admin tool.

It allows users with the SERVER_ADMIN role to modify the roles,
email, etc of users within their organization.  If the user also has
the ORG_ADMIN role, it will allow them to modify any user.

In addition to the role selection, the UI will show what collection of
permissions a particular combination of roles will grant, using tooltips
to explain in more detail.

The user list is searchable by username and can be limited by role.

The organization assignment card will show only the organizations the
active user has permission to modify.  If no organizations are selected,
a Snackbar will pop up explaining that membership in no organizations is
the same as membership in the root (or all) organizations.

In addition to being able to edit users, users may be created using the
same user interface.  Passwords may be set and changed if the server is
set up to use basic authentication.  Otherwise, the password will be
randomized unless explicitly set.