file_store: handle watching artifacts with named sources
When watch_monitoring() is called to monitor an artifact with a named
source, it returns immediately with zero results and nothing in the log.

It turns out that file_store/directory/queue QueuePool.NewListener was
failing silently due to the path containing a slash in it.  This commit
replaces the slash with three dots and reports an error when it
happens in NewListener.