users: extend DeleteUser testcase to ensure org membership was dropped
The TestDeleteUser testcase takes a user which is a member of one org,
adds it to another, removes it from the first, and checks to see if it
still exists and remains in the second org.  It never checks to see if
the membership in the first org was actually completed.