watch/parse_auditd: make it Linux-only
go-libaudit's auparse code assumes that it is running on a Linux system
and may produce strange results on other systems.  It fails to build
entirely on Windows due to its use of the unix package.

Since the likelihood of Mac or Windows clients needing to parse Linux
kernel audit logs is pretty low, let's just make it Linux-only and
avoid it.